# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pâtisserie ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adjustments:
    get:
      operationId: getAdjustments
      tags: [adjustments]
      summary: List adjustments
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: reason
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Adjustments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Adjustment"
    post:
      operationId: createAdjustment
      tags: [adjustments]
      summary: Create adjustment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdjustmentBody"
      responses:
        "201":
          description: Adjustment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adjustment"
  /adjustments/stats:
    get:
      operationId: getAdjustmentsStats
      tags: [adjustments]
      summary: Adjustments loss statistics
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: reason
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Adjustments stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdjustmentsStats"

  # EXPENSES
...
components:
  schemas:
...
    Adjustment:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        productId:
          type: integer
        productName:
          type: string
        quantityChange:
          type: number
        reason:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - reference
        - branchId
        - branchName
        - productId
        - productName
        - quantityChange
        - reason
        - createdAt
    CreateAdjustmentBody:
      type: object
      properties:
        branchId:
          type: integer
        productId:
          type: integer
        quantityChange:
          type: number
        reason:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - branchId
        - productId
        - quantityChange
        - reason
    AdjustmentsStatsReasonRow:
      type: object
      properties:
        reason:
          type: string
        count:
          type: integer
        quantite:
          type: number
        valeur:
          type: number
      required: [reason, count, quantite, valeur]
    AdjustmentsStats:
      type: object
      properties:
        totalPerteQuantite:
          type: number
        totalPerteValeur:
          type: number
        countPertes:
          type: integer
        byReason:
          type: array
          items:
            $ref: "#/components/schemas/AdjustmentsStatsReasonRow"
      required: [totalPerteQuantite, totalPerteValeur, countPertes, byReason]
...